refactoring the folder structure and inserting stg layer manipulation for booking and subscription tables
