+++ lodgify/files/stg/first_subscription_per_id_month.csv
@@ -1,0 +1,3 @@
+subscriber_id,first_subscription_date
+156,2020-01-03
+256,2020-01-01
