+++ lodgify/files/stg/last_booking_per_id_month.csv
@@ -1,0 +1,35 @@
+subscriber_id,booking_year_month,last_booking_date_per_month
+156,2020-01,2020-01-31
+156,2020-02,2020-02-27
+156,2020-03,2020-03-30
+156,2020-04,2020-04-27
+156,2020-05,2020-05-29
+156,2020-06,2020-06-29
+156,2020-07,2020-07-30
+156,2020-08,2020-08-26
+156,2020-09,2020-09-15
+156,2020-10,2020-10-16
+156,2020-11,2020-11-30
+156,2020-12,2020-12-17
+156,2021-01,2021-01-29
+156,2021-03,2021-03-21
+156,2021-04,2021-04-21
+156,2021-05,2021-05-31
+156,2021-06,2021-06-06
+256,2020-01,2020-01-30
+256,2020-02,2020-02-29
+256,2020-03,2020-03-31
+256,2020-04,2020-04-29
+256,2020-05,2020-05-29
+256,2020-06,2020-06-26
+256,2020-07,2020-07-31
+256,2020-08,2020-08-24
+256,2020-09,2020-09-15
+256,2020-10,2020-10-04
+256,2020-12,2020-12-07
+256,2021-01,2021-01-26
+256,2021-02,2021-02-28
+256,2021-03,2021-03-30
+256,2021-04,2021-04-07
+256,2021-05,2021-05-29
+256,2021-06,2021-06-06
