+++ lodgify/files/stg/subscription.csv
@@ -1,0 +1,42 @@
+,id,sub_id,status,datetime_load,sub_year_month
+0,0,156,active,2023-02-25 14:47:09.615238,2020-03
+1,1,156,active,2023-02-25 14:47:09.615238,2020-04
+2,2,156,active,2023-02-25 14:47:09.625033,2020-05
+3,3,156,active,2023-02-25 14:47:09.625951,2020-06
+4,4,156,active,2023-02-25 14:47:09.626622,2020-07
+5,5,156,active,2023-02-25 14:47:09.627270,2020-08
+6,6,156,active,2023-02-25 14:47:09.627912,2020-09
+7,7,156,active,2023-02-25 14:47:09.628489,2020-10
+8,8,156,active,2023-02-25 14:47:09.629126,2020-11
+9,9,156,active,2023-02-25 14:47:09.615238,2020-12
+10,10,156,active,2023-02-25 14:47:09.629895,2021-01
+11,11,156,canceled,2023-02-25 14:47:09.615238,2021-02
+12,12,156,canceled,2023-02-25 14:47:09.630629,2021-03
+13,13,156,active,2023-02-25 14:47:09.615238,2021-04
+14,14,156,active,2023-02-25 14:47:09.631315,2021-05
+15,15,156,active,2023-02-25 14:47:09.632151,2021-06
+16,16,156,active,2023-02-25 14:47:09.615238,2021-07
+17,17,156,active,2023-02-25 14:47:09.615238,2021-08
+18,18,156,,2023-02-25 14:47:09.632866,2021-09
+19,19,156,active,2023-02-25 14:47:09.615238,2021-10
+20,20,156,canceled,2023-02-25 14:47:09.615238,2021-11
+21,21,256,active,2023-02-25 14:47:09.615238,2020-05
+22,22,256,canceled,2023-02-25 14:47:09.615238,2020-06
+23,23,256,active,2023-02-25 14:47:09.615238,2020-07
+24,24,256,canceled,2023-02-25 14:47:09.615238,2020-08
+25,25,256,canceled,2023-02-25 14:47:09.633590,2020-09
+26,26,256,canceled,2023-02-25 14:47:09.634848,2020-10
+27,27,256,canceled,2023-02-25 14:47:09.635778,2020-11
+28,28,256,active,2023-02-25 14:47:09.615238,2020-12
+29,29,256,active,2023-02-25 14:47:09.615238,2021-01
+30,30,256,active,2023-02-25 14:47:09.615238,2021-02
+31,31,256,active,2023-02-25 14:47:09.615238,2021-03
+32,32,256,canceled,2023-02-25 14:47:09.615238,2021-04
+33,33,256,canceled,2023-02-25 14:47:09.615238,2021-05
+34,34,256,,2023-02-25 14:47:09.637176,2021-06
+35,35,256,,2023-02-25 14:47:09.615238,2021-07
+36,36,256,,2023-02-25 14:47:09.637952,2021-08
+37,37,256,active,2023-02-25 14:47:09.615238,2021-09
+38,38,256,active,2023-02-25 14:47:09.615238,2021-10
+39,39,256,active,2023-02-25 14:47:09.615238,2021-11
+40,40,256,canceled,2023-02-25 14:47:09.615238,2021-12
