+++ lodgify/main.py
@@ -1,5 +1,7 @@
-from read_csv import read_csv
+from read_files import read_csv
 
 
 # read_csv.read_bookings_csv()
-read_csv.read_subscription_csv()
+# read_csv.read_subscription_csv()
+# read_csv.sql_query()
+read_csv.stage_layer_manipulation()
