+++ lodgify/union.csv
@@ -1,0 +1,42 @@
+,sub_id,status,dates,datetime_load,formated_dates,next_date,previous_date,previous_status,diff_months,is_duplicate
+16804,156,active,202003,2023-02-25 14:47:09.615238,2020-03-01,2020-04-01,,,1.0,False
+8638,156,active,202004,2023-02-25 14:47:09.615238,2020-04-01,2020-12-01,2020-03-01,active,8.0,False
+0,156,active,,2023-02-25 14:47:09.625033,2020-05-01,,,,,
+0,156,active,,2023-02-25 14:47:09.625951,2020-06-01,,,,,
+0,156,active,,2023-02-25 14:47:09.626622,2020-07-01,,,,,
+0,156,active,,2023-02-25 14:47:09.627270,2020-08-01,,,,,
+0,156,active,,2023-02-25 14:47:09.627912,2020-09-01,,,,,
+0,156,active,,2023-02-25 14:47:09.628489,2020-10-01,,,,,
+0,156,active,,2023-02-25 14:47:09.629126,2020-11-01,,,,,
+879,156,active,202012,2023-02-25 14:47:09.615238,2020-12-01,2021-02-01,2020-04-01,active,2.0,False
+0,156,active,,2023-02-25 14:47:09.629895,2021-01-01,,,,,
+13969,156,canceled,202102,2023-02-25 14:47:09.615238,2021-02-01,2021-04-01,2020-12-01,active,2.0,False
+0,156,canceled,,2023-02-25 14:47:09.630629,2021-03-01,,,,,
+17247,156,active,202104,2023-02-25 14:47:09.615238,2021-04-01,2021-07-01,2021-02-01,canceled,3.0,False
+0,156,active,,2023-02-25 14:47:09.631315,2021-05-01,,,,,
+0,156,active,,2023-02-25 14:47:09.632151,2021-06-01,,,,,
+14735,156,nan,202107,2023-02-25 14:47:09.615238,2021-07-01,2021-08-01,2021-07-01,canceled,1.0,True
+14793,156,nan,202108,2023-02-25 14:47:09.615238,2021-08-01,2021-10-01,2021-08-01,active,2.0,True
+0,156,nan,,2023-02-25 14:47:09.632866,2021-09-01,,,,,
+8542,156,active,202110,2023-02-25 14:47:09.615238,2021-10-01,2021-11-01,2021-08-01,active,1.0,False
+14718,156,canceled,202111,2023-02-25 14:47:09.615238,2021-11-01,2020-05-01,2021-10-01,active,-18.0,False
+10748,256,active,202005,2023-02-25 14:47:09.615238,2020-05-01,2020-06-01,2021-11-01,canceled,1.0,False
+18515,256,canceled,202006,2023-02-25 14:47:09.615238,2020-06-01,2020-07-01,2020-05-01,active,1.0,False
+12968,256,active,202007,2023-02-25 14:47:09.615238,2020-07-01,2020-08-01,2020-06-01,canceled,1.0,False
+232,256,canceled,202008,2023-02-25 14:47:09.615238,2020-08-01,2020-12-01,2020-07-01,active,4.0,False
+0,256,canceled,,2023-02-25 14:47:09.633590,2020-09-01,,,,,
+0,256,canceled,,2023-02-25 14:47:09.634848,2020-10-01,,,,,
+0,256,canceled,,2023-02-25 14:47:09.635778,2020-11-01,,,,,
+2326,256,active,202012,2023-02-25 14:47:09.615238,2020-12-01,2021-01-01,2020-08-01,canceled,1.0,False
+2520,256,nan,202101,2023-02-25 14:47:09.615238,2021-01-01,2021-02-01,2021-01-01,active,1.0,True
+1321,256,nan,202102,2023-02-25 14:47:09.615238,2021-02-01,2021-03-01,2021-02-01,active,1.0,True
+18022,256,nan,202103,2023-02-25 14:47:09.615238,2021-03-01,2021-04-01,2021-03-01,active,1.0,True
+12786,256,canceled,202104,2023-02-25 14:47:09.615238,2021-04-01,2021-05-01,2021-03-01,active,1.0,False
+19597,256,nan,202105,2023-02-25 14:47:09.615238,2021-05-01,2021-07-01,2021-05-01,active,2.0,True
+0,256,nan,,2023-02-25 14:47:09.637176,2021-06-01,,,,,
+12858,256,nan,202107,2023-02-25 14:47:09.615238,2021-07-01,2021-09-01,2021-07-01,active,2.0,True
+0,256,nan,,2023-02-25 14:47:09.637952,2021-08-01,,,,,
+4907,256,active,202109,2023-02-25 14:47:09.615238,2021-09-01,2021-10-01,2021-07-01,active,1.0,False
+12954,256,active,202110,2023-02-25 14:47:09.615238,2021-10-01,2021-11-01,2021-09-01,active,1.0,False
+5903,256,active,202111,2023-02-25 14:47:09.615238,2021-11-01,2021-12-01,2021-10-01,active,1.0,False
+15975,256,canceled,202112,2023-02-25 14:47:09.615238,2021-12-01,,2021-11-01,active,0.0,False
